# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sistema de Gestão Financeira Pessoal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: List savings goals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create goal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get goal by id
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateGoal
      tags: [goals]
      summary: Update goal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete goal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /goals/{id}/deposit:
    post:
      operationId: depositToGoal
      tags: [goals]
      summary: Deposit amount to goal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalDeposit"
      responses:
        "200":
          description: Deposited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

  # Dashboard
...
  /reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Summary for custom period
      security:
        - bearerAuth: []
      parameters:
        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Report summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
  /reports/transactions:
    get:
      operationId: getReportTransactions
      tags: [reports]
      summary: All transactions for a period (for export)
      security:
        - bearerAuth: []
      parameters:
        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Transactions list for period
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  # Admin
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /profile/change-password:
    post:
      operationId: changePassword
      tags: [profile]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Auth schemas
...
    User:
      type: object
      required: [id, name, email, role, currency, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        preferredLang:
          type: string
          enum: [pt, en]
        currency:
          type: string
        createdAt:
          type: string
          format: date-time

    # Account schemas
...
    Transaction:
      type: object
      required: [id, userId, accountId, categoryId, type, amount, date, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        accountId:
          type: integer
        categoryId:
          type: integer
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        description:
          type: ["string", "null"]
        date:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        categoryName:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]
        categoryIcon:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
...
    Goal:
      type: object
      required: [id, userId, name, targetAmount, currentAmount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: ["string", "null"]
        icon:
          type: string
        color:
          type: string
        status:
          type: string
          enum: [active, completed, cancelled]
        progressPercentage:
          type: number
        createdAt:
          type: string
          format: date-time
    GoalInput:
      type: object
      required: [name, targetAmount]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: string
          format: date
        icon:
          type: string
        color:
          type: string
    GoalUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        targetAmount:
          type: number
        deadline:
          type: string
          format: date
        icon:
          type: string
        color:
          type: string
        status:
          type: string
          enum: [active, completed, cancelled]
    GoalDeposit:
      type: object
      required: [amount]
      properties:
        amount:
          type: number

    # Dashboard schemas
...
    CategorySummary:
      type: object
      required: [categoryId, categoryName, total]
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryColor:
          type: ["string", "null"]
        categoryIcon:
          type: ["string", "null"]
        total:
          type: number
...
    ReportSummary:
      type: object
      required: [totalIncome, totalExpense, balance, transactionCount, dateFrom, dateTo]
      properties:
        totalIncome:
          type: number
        totalExpense:
          type: number
        balance:
          type: number
        transactionCount:
          type: integer
        dateFrom:
          type: string
        dateTo:
          type: string
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategorySummary"

    # Admin schemas
...
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        preferredLang:
          type: string
          enum: [pt, en]
        currency:
          type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6